Fix bugs related to connect() method (#46)
* Move new pip deps above py3.12 warning line

* Fix duplicate connect() calls

The connect() method within __init__now calls the
connect() method within the client if one exists.
This enables connection errors to be raised on
connect() instead of on import of stretchpy.

* Fix stall on missing server

When the server is not found, e.g. if the robot is turned off,
then the StretchPy client shouldn't stall on connect(). Instead,
it should report that the server wasn't found after some time.

* Update setup instructions in README