example of simple nav commands
This app provides an example of running simple navigation commands. It is based on the example given at the start of the main Stretch AI documentation in the Quickstart section.

Currently, this code does not run as expected. I am adding it to help us test these basic navigation commands.