handover example
*** WARNING: FOR THIS SIMPLE HANDOVER THE ROBOT HOLDS ITS ARM OUT WITHOUT CHECKING FOR POTENTIAL COLLISIONS. USE AT YOUR OWN RISK ***

This commit provides an example of adding a new task to stretch_ai that can be used with an LLM and speech recognition.

The example task is a simple form of handover that can hand an object to a person. The task first has the robot look around with its head tilted upward. While looking around the robot updates its map. Then, the robot queries the instances in its map to see if there is a person instance. If there is a person instance to which the robot can navigate, the robot navigates to the instance. If successful, the robot lifts its arm, extends its arm, opens its gripper, and retracts its arm. Throughout this process, the robot says what it is attempting to do.

This commit also modifies pickup_prompt.py and pickup_executor.py to enable an LLM to call the handover task when given a natural language request.

Please note that this example task has only been tested a couple of times. The tests took place in a single small room with a single user who used speech to request that the robot bring them a toy chicken sitting near the robot on the floor.

To see how the modified LLM prompt results in calls to the handover task, use the following command.

python -m stretch.app.chat --voice --llm qwen25 --prompt pickup

To try out the hand over task in isolation without the robot holding an object, use the following command.

python -m stretch.app.hand_over_object --target_object "person"

To try out the hand over task together with the other pick and place tasks, use the following command.

python -m stretch.app.ai_pickup --use_llm --use_voice